feat(tts): custom voice preview
WIP for #250

custom preview feature is working

in progress is reorganizing the dialog to look less crowded
