+++ src/renderer/components/SettingsView/index.tsx
@@ -17,7 +17,8 @@ import {
 } from 'shared/data/slices/settings'
 import { TtsEnginesConfigPane } from '../TtsEnginesConfig'
 import { TtsMoreOptionsConfigPane } from '../TtsMoreOptionsConfig'
-import { TtsVoicesConfigPane } from '../TtsVoicesConfig'
+import { TtsBrowseVoicesConfigPane } from '../TtsBrowseVoices'
+import { TtsPreferredVoicesConfigPane } from '../TtsPreferredVoices'
 const { App } = window // The "App" comes from the bridge
 
 enum SelectedMenuItem {
@@ -25,7 +26,8 @@ enum SelectedMenuItem {
     Appearance,
     Behavior,
     Updates,
-    TTSVoices,
+    TTSBrowseVoices,
+    TTSPreferredVoices,
     TTSEngines,
     TTSMoreOptions,
 }
@@ -198,7 +200,7 @@ export function SettingsView() {
                             <li
                                 className={
                                     selectedSection ==
-                                    SelectedMenuItem.TTSVoices
+                                    SelectedMenuItem.TTSBrowseVoices
                                         ? 'selected-menu-item'
                                         : ''
                                 }
@@ -206,11 +208,29 @@ export function SettingsView() {
                                 <button
                                     onClick={(e) =>
                                         setSelectedSection(
-                                            SelectedMenuItem.TTSVoices
+                                            SelectedMenuItem.TTSBrowseVoices
                                         )
                                     }
                                 >
-                                    Voices
+                                    Browse Voices
+                                </button>
+                            </li>
+                            <li
+                                className={
+                                    selectedSection ==
+                                    SelectedMenuItem.TTSPreferredVoices
+                                        ? 'selected-menu-item'
+                                        : ''
+                                }
+                            >
+                                <button
+                                    onClick={(e) =>
+                                        setSelectedSection(
+                                            SelectedMenuItem.TTSPreferredVoices
+                                        )
+                                    }
+                                >
+                                    Preferred Voices
                                 </button>
                             </li>
                             <li
@@ -377,15 +397,36 @@ export function SettingsView() {
                                 checking for updates in the background.
                             </span>
                         </div>
-                    ) : selectedSection == SelectedMenuItem.TTSVoices ? (
+                    ) : selectedSection == SelectedMenuItem.TTSBrowseVoices ? (
                         <div className="tts-voices-config">
                             {pipeline.ttsVoices ? (
-                                <TtsVoicesConfigPane
+                                <TtsBrowseVoicesConfigPane
                                     availableVoices={pipeline.ttsVoices}
                                     userPreferredVoices={
                                         newSettings.ttsConfig.preferredVoices
                                     }
+                                    userDefaultVoices={
+                                        newSettings.ttsConfig.defaultVoices
+                                    }
+                                    onChangePreferredVoices={
+                                        onTtsVoicesPreferenceChange
+                                    }
+                                    onChangeDefaultVoices={
+                                        onTtsVoicesDefaultsChange
+                                    }
+                                />
+                            ) : (
+                                <p>Loading voices...</p>
+                            )}
+                        </div>
+                    ) : selectedSection == SelectedMenuItem.TTSPreferredVoices ? (
+                        <div className="tts-voices-config">
+                            {pipeline.ttsVoices ? (
+                                <TtsPreferredVoicesConfigPane
                                     ttsEnginesStates={pipeline.ttsEnginesStates}
+                                    userPreferredVoices={
+                                        newSettings.ttsConfig.preferredVoices
+                                    }
                                     userDefaultVoices={
                                         newSettings.ttsConfig.defaultVoices
                                     }
