+++ src/renderer/components/TtsEnginesConfig/index.tsx
@@ -152,9 +152,9 @@ export function TtsEnginesConfigPane({
     console.log(engineProperties)
     return (
         <>
-            <p id="available-voices-label" className="label">
-                <b>Configure text-to-speech engines</b>
-            </p>
+            <h2 id="configure-tts-label" className="label">
+                Configure text-to-speech engines
+            </h2>
             <p className="desc">
                 After configuring these engines with the required credentials,
                 they will be available under 'Voices'.
